requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
openai>=1.0.0
openai-whisper>=20231117
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
nltk>=3.8.1
scikit-learn>=1.3.0
textblob>=0.17.1
sumy>=0.11.0
numpy>=1.24.0